feature: add README and deployment instructions
