+++ package.json
@@ -60,6 +60,7 @@
     "eslint-config-standard-with-typescript": "^23.0.0",
     "eslint-plugin-prettier": "^4.2.1",
     "prettier": "^2.8.0",
+    "rimraf": "^4.1.2",
     "rollup": "^3.10.0",
     "rollup-plugin-dts": "^5.1.1",
     "rollup-plugin-esbuild": "^5.0.0",
@@ -71,6 +72,7 @@
   },
   "dependencies": {
     "dotenv": "^16.0.3",
+    "fastify": "^4.12.0",
     "node-telegram-bot-api": "^0.61.0"
   }
 }
